==> analysis/01-batch-integration.Rmd <==
---
title: "Batch integration"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Motivation

## Setup

Setup chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/01-batch-integration/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE,
  fig.align = "center",
  out.width = '100%'
)
```

Load Bioconductor packages:

```{r}
library(batchelor)
library(EnsDb.Mmusculus.v79)
library(scater)
library(scran)
library(scuttle)
```

Load CRAN packages:

```{r}
library(ggforce)
library(pheatmap)
```

Source user functions:

```{r}
source("code/splitByCol.R")
```

## Data

### Test data

Read experiment objects:

```{r}
sce <- list(
  NMP_EHF = readRDS("data/NMP_EHF.rds"),
  LPMP_EHF = readRDS("data/LPMP_EHF.rds"),
  LPMP_LHF = readRDS("data/LPMP_LHF.rds")
)
```

## Integration

### Data processing

Define function to clean SCE object:

```{r}
cleanSCE <- function(x) {

  # Clear metadata
  metadata(x) <- list()

  # Replace rownames
  rownames(x) <- rowData(x)$gene_id

  # Clear rowData
  rowData(x) <- NULL

  # Replace rowRanges
  rowRanges(x) <- relist(GRanges(), PartitioningByEnd(integer(length(x)), names = rownames(x)))

  # Clear colnames
  colnames(x) <- x$sample
  
  # Replace colData
  colData(x)$cluster <- paste(x$region, x$stage, x$cluster, sep = "_")

  # Clear reducedDims
  reducedDims(x) <- NULL
  
  # Return object
  return(x)

}
```

Clean experiment objects:

```{r}
sce <- lapply(sce, cleanSCE)
```

Subset data to common features:

```{r}
ids <- Reduce(intersect, lapply(sce, rownames))

sce <- lapply(sce, "[", i = ids)
```

### Variance modelling

Model the per-gene variance:

```{r}
dec <- lapply(sce, function(x) modelGeneVarWithSpikes(x, spikes = "ERCC", block = x$phase))

#dec <- lapply(sce, modelGeneVarWithSpikes, "ERCC")
```

Combine variance decompositions:

```{r}
dec <- do.call(combineVar, dec)
```

### Feature selection

Import genes to ignore for feature selection

```{r}
nan <- read.csv("data/blacklist.csv")

dec$blacklist <- rownames(dec) %in% nan$ENSEMBL
```

Identify highly variable genes:

```{r}
hvg <- subset(dec, blacklist == FALSE)

hvg <- getTopHVGs(dec, var.field = "bio", var.threshold = 0, fdr.field = "FDR", fdr.threshold = 0.05)
```

### Batch normalization

Perform scaling normalization:

```{r}
sce <- multiBatchNorm(sce)
```

### Batch integration

Split each batch by phase:

```{r}
sce <- lapply(sce, function(x) splitByCol(x, x$phase))
```

Flatten nested list of objects:

```{r}
sce <- lapply(rapply(sce, enquote, how = "unlist"), eval)
```

Correct experiments using nearest neighbours:

```{r}
set.seed(0001011110)

mnn <- correctExperiments(
  sce,
  subset.row = hvg,
  correct.all = TRUE,
  PARAM = FastMnnParam(auto.merge = TRUE, get.variance = TRUE)
)
```

Store combined variance model:

```{r}
metadata(mnn)$combineVar <- dec
```

Store highly variable genes:

```{r}
metadata(mnn)$getTopHVGs <- hvg
```

Replace rowData slot:

```{r}
rowData(mnn)$gene_id <- rownames(mnn)

rowData(mnn)$gene_name <- mapIds(EnsDb.Mmusculus.v79, keys = rownames(mnn), column = "GENENAME", keytype = "GENEID")
```

## Diagnostics

Check proportion of lost variance:

```{r}
var <- metadata(mnn)$merge.info$lost.var

col <- colorRampPalette(RColorBrewer::brewer.pal(n = 5, name = "Reds"))(100)

brk <- seq(0, max(var, na.rm = TRUE), length.out = 101)

pheatmap(var, color = col, display_numbers = TRUE, number_format = "%.2f", cluster_cols = FALSE, cluster_rows = FALSE, angle_col = 0)
```

Cluster on the corrected PCA to obtain a partitioning of the cells:

```{r}
snn <- buildSNNGraph(mnn, type = "jaccard", use.dimred = "corrected")

com <- igraph::cluster_louvain(snn)
```

Tabulate number of cells from multiple batches within each cluster:

```{r}
tab <- table(Cluster = com$membership, Batch = mnn$batch)

col <- colorRampPalette(RColorBrewer::brewer.pal(n = 5, name = "Blues"))(100)

brk <- seq(0, max(tab), length.out = 101)

pheatmap(tab, color = col, display_numbers = TRUE, number_format = "%d", cluster_cols = FALSE, cluster_rows = FALSE, angle_col = 0)
```

Visualize the corrected PCA:

```{r}
ggcells(mnn, aes(corrected.1, corrected.2, colour = cluster)) + 
  geom_point(size = 1) + 
  scale_colour_brewer(name = "Cluster", palette = "Set3") + 
  labs(x = "PCA 1", y = "PCA 2") + 
  facet_wrap(~ batch) + 
  theme_no_axes() + 
  coord_fixed()
```

Perform TSNE on the corrected PCA matrix:

```{r}
set.seed(0000110000)

mnn <- runTSNE(mnn, dimred = "corrected")
```

Visualize the corrected PCA using a TSNE plot:

```{r}
ggcells(mnn, aes(TSNE.1, TSNE.2, colour = cluster)) +
  geom_point(size = 1) +
  scale_colour_brewer(name = "Cluster", palette = "Set3") + 
  labs(x = "TSNE 1", y = "TSNE 2") +
  facet_wrap(~ batch) +
  theme_no_axes() + 
  theme(aspect.ratio = 1)
```

Perform UMAP on the corrected PCA matrix:

```{r}
set.seed(1100001001)

mnn <- runUMAP(mnn, dimred = "corrected")
```

Visualize the corrected PCA using a UMAP plot:

```{r}
ggcells(mnn, aes(UMAP.1, UMAP.2, colour = cluster)) +
  geom_point(size = 1) +
  scale_colour_brewer(name = "Cluster", palette = "Set3") +
  labs(x = "UMAP 1", y = "UMAP 2") +
  facet_wrap(~ batch) +
  theme_no_axes() + 
  theme(aspect.ratio = 1)
```

## Summary

### Output

Save experiment object to disk:

```{r}
saveRDS(mnn, file = "output/01-batch-integration.rds")
```

### Session

Print session information:

```{r}
sessionInfo()
```
